Andrew updated the repo licence
The Apache License was chosen as it is a permissive free software license written by the Apache Software Foundation. It allows users to use the software for any purpose, to distribute it, to modify it, and to distribute modified versions of the software under the terms of the license, without concern for royalties. It's goal is to provide reliable and long-lived software products through collaborative open source software development and as it is a non-profit organisation all proceeds go back into software development which I thought was a cool aspect